backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.3.2
openai==1.3.0
requests==2.31.0
Pillow==10.1.0
numpy==1.25.2
nltk==3.8.1
werkzeug==2.3.7
textstat==0.7.3
